cat server.py | GPT_MODEL=gpt-4 PROMPT_PREFIX='Why do I get a 404 when I curl localhost:5000/ with this flask server?  ' chatgpt
The reason you are getting a 404 error when curling localhost:5000/ is that the index.html file is not being served properly. In the index() function, app.send_static_file('index.html') is used to serve the index.html file but Flask needs to know where to find the static files. You can add the following line to the app to tell Flask where to look for the static files:

app = Flask(__name__, static_url_path='', static_folder='static')

This tells Flask to look for the static files in the 'static' directory in the same directory as the Flask server script. After adding this line, Flask will be able to serve the index.html file correctly and you should no longer receive a 404 error when curling localhost:5000/.